Index created contacts via post-commit hook to avoid duplicate ES writes
